refactor to data folder for blogs, projects, about-me
