+++ front-end/src/components/home/AboutMeSession.js
@@ -1,3 +1,5 @@
+import data from "./../../data/about-me.json";
+
 import TitleSession from "../others/TitleSession";
 import style from "./style/AboutMeSession.module.css";
 
@@ -8,18 +10,7 @@ const AboutMeSession = () => {
 
       <div className={style.aboutMeSession}>
         <div className={style.descriptionArea}>
-          <div className={style.description}>
-            <p>Hello, I'm Long!</p>
-            <p> I'm a passionate backend developer from Ho Chi Minh
-            City Open University (OU HCM), with a background in Information
-            Technology. I specialize in building scalable and secure backend
-            systems.</p>
-            <p> With hands-on experience in backend frameworks and database
-            management, I am dedicated to continuous learning and staying
-            updated with the latest technologies. I have a strong
-            problem-solving mindset and a collaborative approach to working in
-            development teams.</p>
-          </div>
+          <div className={style.description}>{data.content}</div>
 
           <div className={style.btnReadMore}>Read more</div>
         </div>
