+++ front-end/src/data/projects.json
