+++ front-end/src/pages/blog/BlogListPage.js
@@ -1,3 +1,5 @@
+import data from "./../../data/blogs.json";
+
 import { BLOGS_PER_PAGE } from "./../../config/pagination";
 import { useState } from "react";
 
@@ -5,8 +7,6 @@ import { useState } from "react";
 import BlogPaginator from "../../components/blog/BlogPaginator";
 import PageDescription from "../../components/others/PageDescription";
 import TitleSession from "../../components/others/TitleSession";
-
-import data from "./data/data";
 
 const BlogListPage = () => {
   const [description] = useState(data.pageDescription);
